+++ src/models/response/add-card-form.model.ts
@@ -7,18 +7,13 @@ export class AddCardFormResponse extends Response {
   /**
    * Data response.
    */
-  data: AddCardFormData
+  data?: AddCardFormData
 }
 
 /**
  * Class AddCardFormData
  */
 export class AddCardFormData {
-  /**
-   * Assigned transaction ID for the payment.
-   */
-  transactionId: string
-
   /**
    * URL to hosted payment gateway.
    */