+++ src/models/response/request-settlements.model.ts
@@ -7,7 +7,7 @@ export class SettlementsResponse extends Response {
   /**
    * Data response.
    */
-  data: SettlementsData
+  data?: SettlementsData
 }
 
 /**