+++ src/utils/handle-request.util.ts
@@ -1,3 +1,6 @@
-export const handleRequest = <T>(promise: Promise<T>) => {
-  return promise.then((data) => [undefined, data] as const).catch((err) => [err, undefined] as const)
+export const handleRequest = async <T>(promise: Promise<T>): Promise<readonly [any, T]> => {
+  return promise
+    .then((data): readonly [any, T] => [undefined, data] as const)
+    .catch((err): readonly [any, T] => [err, undefined] as const)
+
 }