+++ CHANGELOG.md
@@ -4,23 +4,43 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.4] - 2025-06-05
+
+### Added
+
+- Network token to `GetTokenResponse`
+- Github workflow file
+
+### Fixed
+
+- Fixed `createAddCardFormRequest` to return card form url
+
 ## [1.0.3] - 2024-10-01
+
 ### Changed
+
 - Updated VAT rate decimal support
 - Updated tests
 
 ### Fixed
+
 - Fixed validation for `CreatePaymentRequest` groups
 
 ## [1.0.2] - 2024-03-22
+
 ### Changed
+
 - Bump follow-redirects
 
 ## [1.0.1] - 2024-02-22
+
 ### Changed
+
 - Moved package to npm public repository
 - Bumped nodejs and follow-redirects versions
 
 ## [1.0.0] - 2023-11-28
+
 ### Added
+
 - Initial release